+++ test.js
@@ -102,6 +102,24 @@ test('Test always consumes token from response', async t => {
   t.is(client.options.token, 'foobar')
 })
 
+test('Test removes token on 401 responses', async t => {
+  const client = new ScreepsModules({
+    token: 'foobar'
+  })
+
+  nock('https://screeps.com')
+    .get('/unauthorized')
+    .reply(401, 'Unauthorized')
+
+  try {
+    await client.request('/unauthorized')
+
+    t.fail()
+  } catch (e) {
+    t.is(client.options.token, '')
+  }
+})
+
 test('Test custom server URL', async t => {
   t.plan(1)